Deep-Live-Cam/requirements.txt
numpy>=1.23.5,<2
requests>=2.28.0
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.18.0
insightface==0.7.3
psutil==5.9.8
tk==0.1.0
customtkinter==5.2.2
pillow==12.1.1
tqdm>=4.65.0
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.23.2; sys_platform != 'darwin'
tensorflow>=2.15.0; sys_platform != 'darwin'
tensorflow>=2.15.0; sys_platform == 'darwin' and python_version < '3.13'
opennsfw2==0.10.2
protobuf==4.25.1
pygrabber; sys_platform == 'win32'
pyvirtualcam>=0.11.0
